Replace gpgme with NaCl's SecretBox
GPGME doesn't work very well with gpg2. Instead, use NaCl's SecretBox for
authenticated encryption.

The one interesting modification to vanilla SecretBox (or, rather,
RandomNonceBox) is that we've placed PBKDF2 with a configurable number of rounds
between the user's password and what's actually handed off to SecretBox. The
primary benefit is that the user can provide a key of whatever length he or she
desires, and doesn't need to know anything about the underlying key size.

This modification is probably "pretty secure," in the sense that NaCl is "pretty
secure" and that we're using a standard password derivation mechanism using a
"pretty secure" hash function (sha256) in a very standard way. However, nobody
who Knows What They Are Doing (tm) has reviewed this code, so keep that in mind
when evaluating its security.